Factoring out common view code
